Only block on community/home/all feeds
