+++ src/features/feed/PostCommentFeed.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useContext, useEffect, useRef } from "react";
 import Feed, { FeedProps, FetchFn } from "./Feed";
 import FeedComment from "../comment/inFeed/FeedComment";
-import { CommentView, Post as PostType, PostView } from "lemmy-js-client";
+import { CommentView, PostView } from "lemmy-js-client";
 import { useAppDispatch, useAppSelector } from "../../store";
 import { css } from "@emotion/react";
 import { postHiddenByIdSelector, receivedPosts } from "../post/postSlice";
@@ -9,6 +9,7 @@ import { receivedComments } from "../comment/commentSlice";
 import Post from "../post/inFeed/Post";
 import CommentHr from "../comment/CommentHr";
 import { FeedContext } from "./FeedContext";
+import { postHasFilteredKeywords } from "../../helpers/lemmy";
 
 const thickBorderCss = css`
   border-bottom: 8px solid var(--thick-separator-color);
@@ -28,12 +29,14 @@ interface PostCommentFeed
   extends Omit<FeedProps<PostCommentItem>, "renderItemContent"> {
   communityName?: string;
   filterHiddenPosts?: boolean;
+  filterKeywords?: boolean;
 }
 
 export default function PostCommentFeed({
   communityName,
   fetchFn: _fetchFn,
   filterHiddenPosts = true,
+  filterKeywords = true,
   ...rest
 }: PostCommentFeed) {
   const dispatch = useAppDispatch();
@@ -112,8 +115,11 @@ export default function PostCommentFeed({
   const filterFn = useCallback(
     (item: PostCommentItem) =>
       !postHiddenById[item.post.id] &&
-      !postHasFilteredKeywords(item.post, filteredKeywords),
-    [postHiddenById, filteredKeywords],
+      !postHasFilteredKeywords(
+        item.post,
+        filterKeywords ? filteredKeywords : [],
+      ),
+    [postHiddenById, filteredKeywords, filterKeywords],
   );
 
   const getIndex = useCallback(
@@ -132,20 +138,4 @@ export default function PostCommentFeed({
       itemsRef={itemsRef}
     />
   );
-}
-
-function postHasFilteredKeywords(post: PostType, keywords: string[]): boolean {
-  for (const keyword of keywords) {
-    if (keywordFoundInSentence(keyword, post.name)) return true;
-  }
-
-  return false;
-}
-
-function keywordFoundInSentence(keyword: string, sentence: string): boolean {
-  // Create a regular expression pattern to match the keyword as a whole word
-  const pattern = new RegExp(`\\b${keyword}\\b`, "i");
-
-  // Use the RegExp test method to check if the pattern is found in the sentence
-  return pattern.test(sentence);
 }
