+++ src/features/user/Profile.tsx
@@ -114,6 +114,7 @@ export default function Profile({ person }: ProfileProps) {
       fetchFn={fetchFn}
       header={header}
       filterHiddenPosts={false}
+      filterKeywords={false}
     />
   );
 }
